services/email-worker/requirements.txt
# FastAPI and ASGI server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Email handling
fastapi-mail>=0.0.0
aiosmtplib>=0.0.1

# Dapr integration
dapr>=1.14.0
dapr-ext-fastapi>=1.14.0

# Environment variables
pydantic-settings>=2.6.1
python-dotenv==1.0.0

# Database (for reading task data)
sqlmodel>=0.0.19
asyncpg>=0.29.0
pydantic>=2.8.0

# Logging
python-json-logger>=2.0.7

# HTML templates
jinja2>=3.1.2

# Utilities
python-multipart==0.0.6
